requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
pytest==8.2.0
pytest-cov==5.0.0
bcrypt==5.0.0